test: split classifier routing assertions
